+++ fsm-core/src/test/kotlin/io/github/ngirchev/fsm/it/ExDomainFsmIT.kt
@@ -58,15 +58,15 @@ internal class ExDomainFsmIT {
             .add(from = AUTO_SENT, onEvent = "TO_END", to = DONE)
             .build()
         val table2 = ExTransitionTable.Builder<DocumentState, String>()
-            .from(NEW).to(READY_FOR_SIGN).onEvent("TO_READY").end()
+            .from(NEW).onEvent("TO_READY").to(READY_FOR_SIGN).end()
             .from(READY_FOR_SIGN).toMultiple()
-            .to(SIGNED).onEvent("USER_SIGN").timeout(Timeout(1)).end()
-            .to(CANCELED).onEvent("FAILED_EVENT").end().endMultiple()
+            .onEvent("USER_SIGN").to(SIGNED).timeout(Timeout(1)).end()
+            .onEvent("FAILED_EVENT").to(CANCELED).end().endMultiple()
             .from(SIGNED).onEvent("FAILED_EVENT").to(CANCELED).end()
-            .from(SIGNED).onEvent("TO_END").toMultiple()                  // switch case example
-            .to(AUTO_SENT).onCondition(signRequiredGuard).end()            // first
-            .to(DONE).onCondition(notSignRequiredGuard).end()             // second
-            .to(CANCELED).end().endMultiple()                           // else
+            .from(SIGNED).toMultiple()                  // switch case example
+            .onEvent("TO_END").to(AUTO_SENT).onCondition(signRequiredGuard).end()            // first
+            .onEvent("TO_END").to(DONE).onCondition(notSignRequiredGuard).end()             // second
+            .onEvent("TO_END").to(CANCELED).end().endMultiple()                           // else
             .from(AUTO_SENT).onEvent("TO_END").to(DONE).end()
             .build()